Refractored Code

+++ src/WorkSpace.java
@@ -1,8 +1,8 @@
 public class WorkSpace implements Space, UsableSpace {
-    double workStations;
-    public WorkSpace(double workStations){
+    int workStations;
+    public WorkSpace(int workStations){
         this.workStations = workStations;
     }
 
-    public double workStations(){return this.workStations;}
+    public int workStations(){return this.workStations;}
 }
